Adding notification feature for comments, likes and new followers

+++ src/app/api/posts/[postId]/likes/route.ts
@@ -59,6 +59,17 @@ export async function POST(
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
 
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true,
+      },
+    });
+
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
+    }
+
     await prisma.$transaction([
       prisma.like.upsert({
         where: {
@@ -73,6 +84,18 @@ export async function POST(
         },
         update: {},
       }),
+      ...(loggedInUser.id !== post.userId
+        ? [
+            prisma.notification.create({
+              data: {
+                issuerId: loggedInUser.id,
+                recipientId: post.userId,
+                postId,
+                type: "LIKE",
+              },
+            }),
+          ]
+        : []),
     ]);
 
     return new Response();
@@ -93,6 +116,17 @@ export async function DELETE(
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
 
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true,
+      },
+    });
+
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
+    }
+
     await prisma.$transaction([
       prisma.like.deleteMany({
         where: {
@@ -100,6 +134,14 @@ export async function DELETE(
           postId,
         },
       }),
+      prisma.notification.deleteMany({
+        where: {
+          issuerId: loggedInUser.id,
+          recipientId: post.userId,
+          postId,
+          type: "LIKE",
+        }
+      })
     ]);
 
     return new Response();
